move tests into test_array:
